backend/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
codecov==2.1.13
coverage==7.6.8
Django==4.2.16
django-apps==0.1.0
django-cors-headers==4.6.0
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2
elastic-transport==8.15.1
elasticsearch==8.17.0
elasticsearch-dsl==8.17.0
flake8==7.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
confluent_kafka==2.8.0
mccabe==0.7.0
msgpack==1.1.0
multimethod==1.12
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pynvim==0.5.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
